payment-handler/can-make-payment-event.https.html

<!doctype html>
<meta charset="utf-8">
<title>Tests for CanMakePaymentEvent</title>
<link rel="help" href="https://w3c.github.io/payment-handler/#the-canmakepaymentevent">
<link rel="manifest" href="manifest.json">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async function registerAppAndBuildPaymentRequest(methodName) {
  await navigator.serviceWorker.register("app-can-make-payment.js");
  const registration = await navigator.serviceWorker.ready;
  if (!registration.paymentManager) {
    return;
  }
  if (registration.paymentManager.requestPermission) {
    const permission = await registration.paymentManager.requestPermission();
    if (permission !== 'granted') {
      return;
    }
  }
  await registration.paymentManager.instruments.set("instrument-key", {
    name: "Test Payment Method",
    enabledMethods: [methodName],
  });
  return new PaymentRequest(
    [
      {
        supportedMethods: methodName,
      },
    ],
    {
      total: {
        label: "Total",
        amount: {
          currency: "USD",
          value: "0",
        },
      },
    }
  );
}

promise_test(async t => {
  const methodName = window.location.origin + "/canMakePayment-false";
  const request = await registerAppAndBuildPaymentRequest(methodName);
  assert_not_equals(request, undefined);
  let paymentRequestCanMakePaymentResult;
  try {
    paymentRequestCanMakePaymentResult = await request.canMakePayment();
  } catch (err) {
    assert_equals(
      err.name,
      "NotAllowedError",
      "If it throws, then it must be NotAllowedError"
    );
  }
  assert_false(paymentRequestCanMakePaymentResult, "canMakePayment() must return false.");
  await promise_rejects(t, "NotSupportedError", request.show());
}, "If CanMakePaymentEvent.respondWith(false) is called, then the payment method is not supported.");

promise_test(async t => {
  const methodName = window.location.origin + "/canMakePayment-promise-false";
  const request = await registerAppAndBuildPaymentRequest(methodName);
  assert_not_equals(request, undefined);
  let paymentRequestCanMakePaymentResult;
  try {
    paymentRequestCanMakePaymentResult = await request.canMakePayment();
  } catch (err) {
    assert_equals(
      err.name,
      "NotAllowedError",
      "If it throws, then it must be NotAllowedError"
    );
  }
  assert_false(paymentRequestCanMakePaymentResult, "canMakePayment() must return false.");
  await promise_rejects(t, "NotSupportedError", request.show());
}, "If CanMakePaymentEvent.respondWith(Promise.resolve(false)) is called, then the payment method is not supported.");

promise_test(async t => {
  const methodName = window.location.origin + "/canMakePayment-true";
  const request = await registerAppAndBuildPaymentRequest(methodName);
  assert_not_equals(request, undefined);
  let paymentRequestCanMakePaymentResult;
  try {
    paymentRequestCanMakePaymentResult = await request.canMakePayment();
  } catch (err) {
    assert_equals(
      err.name,
      "NotAllowedError",
      "If it throws, then it must be NotAllowedError"
    );
  }
  assert_true(paymentRequestCanMakePaymentResult, "canMakePayment() must return true.");
  const acceptPromise = request.show();
  await request.abort();
  await promise_rejects(t, "AbortError", acceptPromise);
}, "If CanMakePaymentEvent.respondWith(true) is called, then the payment method is supported.");

promise_test(async t => {
  const methodName = window.location.origin + "/canMakePayment-promise-true";
  const request = await registerAppAndBuildPaymentRequest(methodName);
  assert_not_equals(request, undefined);
  let paymentRequestCanMakePaymentResult;
  try {
    paymentRequestCanMakePaymentResult = await request.canMakePayment();
  } catch (err) {
    assert_equals(
      err.name,
      "NotAllowedError",
      "If it throws, then it must be NotAllowedError"
    );
  }
  assert_true(paymentRequestCanMakePaymentResult, "canMakePayment() must return true.");
  const acceptPromise = request.show();
  await request.abort();
  await promise_rejects(t, "AbortError", acceptPromise);
}, "If CanMakePaymentEvent.respondWith(Promise.resolve(true)) is called, then the payment method is supported.");

promise_test(async t => {
  const methodName = window.location.origin + "/canMakePayment-custom-error";
  const request = await registerAppAndBuildPaymentRequest(methodName);
  assert_not_equals(request, undefined);
  let paymentRequestCanMakePaymentResult;
  try {
    paymentRequestCanMakePaymentResult = await request.canMakePayment();
  } catch (err) {
    assert_equals(
      err.name,
      "NotAllowedError",
      "If it throws, then it must be NotAllowedError"
    );
  }
  assert_false(paymentRequestCanMakePaymentResult, "canMakePayment() must return false.");
  await promise_rejects(t, "NotSupportedError", request.show());
}, "If CanMakePaymentEvent.respondWith(Promise.reject(error)) is called, then the payment method is not supported.");

</script>
